Automate the SonarQube analysis with bazel
This change enables automation of SonarQube analysis for the heartbeat
plugin. This plugin could now run the sonar by calling the sonar script.

Note that this change should not be merged until change 175010 is merged
and the SHA1 of the load_bazlets in the WORKSPACE file is updated to be
the one of the merged change 175010.

Change-Id: Iaae8820096feeb58769d640aca44d25eed2c9da5